% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqsample-package.R
\docType{package}
\name{dqsample-package}
\alias{dqsample}
\alias{dqsample-package}
\title{dqsample: Bias-Free Alternative to R's 'sample' Function}
\description{
The 'base::sample' function uses a slightly biased algorithm for
  creating random integers within a given range. As an alternative the
  algorithm suggested by Daniel Lemire (2018, <arXiv:1805.1094>) is used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.daqana.org/dqsample}
  \item \url{https://github.com/daqana/dqsample}
  \item Report bugs at \url{https://github.com/daqana/dqsample/issues}
}

}
